Added authorization.

+++ backend/routes/posts.js
@@ -37,7 +37,8 @@ router.post('', checkAuth, multer({ storage: storage }).single('image'), (req, r
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
 
   post.save().then(createdPost => {
@@ -61,12 +62,17 @@ router.patch('/:id', checkAuth, multer({ storage: storage }).single('image'), (r
     _id: req.params.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
 
-  Post.updateOne({ _id: req.params.id }, post)
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
     .then(result => {
-      res.status(200).json({ message: 'Update successfull!' });
+      if (result.nModified > 0) {
+        res.status(200).json({ message: 'Update successfull!' });
+      } else {
+        res.status(401).json({ message: 'Not authorized!' });
+      }
     });
 });
 
@@ -107,9 +113,13 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then(result => {
-      res.status(200).json({ message: 'Post is deleted!' });
+      if (result.n > 0) {
+        res.status(200).json({ message: 'Post is deleted!' });
+      } else {
+        res.status(401).json({ message: 'Not authorized!' });
+      }
     });
 });
